refactor: move SceneBoundsChecker interfaces to their own assemblies
- interfaces were moved from the original files to new ones under `Interfaces` folder.
- used that folder to create new assembly
- fixed all assembly definition errors
