+++ unity-renderer/Assets/Builder/Scripts/Builder.asmdef
@@ -24,7 +24,8 @@
     "GUID:6dacc6cde5056a34c9488c309ca0a18a",
     "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
