+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Font/Tests/FontTests.asmdef
@@ -1,33 +1,34 @@
 {
-    "name": "FontTests",
-    "rootNamespace": "",
-    "references": [
-        "GUID:e555144732b726c4cba5b0f6337fea75",
-        "GUID:b1087c5731ff68448a0a9c625bb7e52d",
-        "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
-        "GUID:6055be8ebefd69e48b49212b09b47b2f",
-        "GUID:27619889b8ba8c24980f49ee34dbb44a",
-        "GUID:0acc523941302664db1f4e527237feb3",
-        "GUID:a881b57670b1d2747a6d7f9e32b63230",
-        "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
-        "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
-        "GUID:86e5747efa9128344bd8e4754db51d3b",
-        "GUID:d7ced3c8f2800ed40a132e6df054755e",
-        "GUID:1dd0780aa6be12b428c8005d0bee46b8",
-        "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:800ec6f2bae94e64098ee2c3a5e5d818"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [],
-    "noEngineReferences": false
+  "name": "FontTests",
+  "rootNamespace": "",
+  "references": [
+    "GUID:e555144732b726c4cba5b0f6337fea75",
+    "GUID:b1087c5731ff68448a0a9c625bb7e52d",
+    "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
+    "GUID:6055be8ebefd69e48b49212b09b47b2f",
+    "GUID:27619889b8ba8c24980f49ee34dbb44a",
+    "GUID:0acc523941302664db1f4e527237feb3",
+    "GUID:a881b57670b1d2747a6d7f9e32b63230",
+    "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
+    "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
+    "GUID:86e5747efa9128344bd8e4754db51d3b",
+    "GUID:d7ced3c8f2800ed40a132e6df054755e",
+    "GUID:1dd0780aa6be12b428c8005d0bee46b8",
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": true,
+  "precompiledReferences": [
+    "nunit.framework.dll"
+  ],
+  "autoReferenced": false,
+  "defineConstraints": [
+    "UNITY_INCLUDE_TESTS"
+  ],
+  "versionDefines": [],
+  "noEngineReferences": false
 }
