+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/Materials/Tests/MaterialsTests.asmdef
@@ -11,7 +11,8 @@
     "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
     "GUID:b308ae946900fa04080222d7daf47d9a",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
