+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/QuestTrackingInfo/Tests/QuestTrackingInfoTests.asmdef
@@ -1,34 +1,35 @@
 {
-    "name": "QuestTrackingInfoTests",
-    "rootNamespace": "",
-    "references": [
-        "GUID:e555144732b726c4cba5b0f6337fea75",
-        "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
-        "GUID:81a6b1347b86bfe4aa02ed089ff533c1",
-        "GUID:27619889b8ba8c24980f49ee34dbb44a",
-        "GUID:86e5747efa9128344bd8e4754db51d3b",
-        "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
-        "GUID:d7ced3c8f2800ed40a132e6df054755e",
-        "GUID:a881b57670b1d2747a6d7f9e32b63230",
-        "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
-        "GUID:9b06d63e1a13bc34c995683595d8d849",
-        "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
-        "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:800ec6f2bae94e64098ee2c3a5e5d818"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "NSubstitute.dll",
-        "nunit.framework.dll",
-        "System.Threading.Tasks.Extensions.dll"
-    ],
-    "autoReferenced": true,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [],
-    "noEngineReferences": false
+  "name": "QuestTrackingInfoTests",
+  "rootNamespace": "",
+  "references": [
+    "GUID:e555144732b726c4cba5b0f6337fea75",
+    "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
+    "GUID:81a6b1347b86bfe4aa02ed089ff533c1",
+    "GUID:27619889b8ba8c24980f49ee34dbb44a",
+    "GUID:86e5747efa9128344bd8e4754db51d3b",
+    "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
+    "GUID:d7ced3c8f2800ed40a132e6df054755e",
+    "GUID:a881b57670b1d2747a6d7f9e32b63230",
+    "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
+    "GUID:9b06d63e1a13bc34c995683595d8d849",
+    "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": true,
+  "precompiledReferences": [
+    "NSubstitute.dll",
+    "nunit.framework.dll",
+    "System.Threading.Tasks.Extensions.dll"
+  ],
+  "autoReferenced": true,
+  "defineConstraints": [
+    "UNITY_INCLUDE_TESTS"
+  ],
+  "versionDefines": [],
+  "noEngineReferences": false
 }
