+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Components/UUIDComponent/Tests/UUIDComponentTests.asmdef
@@ -15,17 +15,18 @@
         "GUID:0acc523941302664db1f4e527237feb3",
         "GUID:a881b57670b1d2747a6d7f9e32b63230",
         "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
-        "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
-        "GUID:1dd0780aa6be12b428c8005d0bee46b8",
-        "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
-        "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
-        "GUID:86e5747efa9128344bd8e4754db51d3b",
-        "GUID:d7ced3c8f2800ed40a132e6df054755e",
-        "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
-        "GUID:cc6bfabbfe87b7949aa248893f89ce37",
-        "GUID:b308ae946900fa04080222d7daf47d9a",
-        "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:800ec6f2bae94e64098ee2c3a5e5d818"
+      "GUID:a78c5d8dad39efe45ab1b21901f8db0f",
+      "GUID:1dd0780aa6be12b428c8005d0bee46b8",
+      "GUID:1320d33ea2522ba4e8e60cf9e6a351dd",
+      "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
+      "GUID:86e5747efa9128344bd8e4754db51d3b",
+      "GUID:d7ced3c8f2800ed40a132e6df054755e",
+      "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
+      "GUID:cc6bfabbfe87b7949aa248893f89ce37",
+      "GUID:b308ae946900fa04080222d7daf47d9a",
+      "GUID:97d8897529779cb49bebd400c7f402a6",
+      "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+      "GUID:5289a39f46259954ca74010d7c9ded16"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
