+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/BuilderMode/BuildingTool.asmdef
@@ -50,7 +50,8 @@
     "GUID:28f74c468a54948bfa9f625c5d428f56",
     "GUID:e99faafecbf8dd74d8130597511b9b27",
     "GUID:6e6a87f97192ec947993f360104e75b7",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
