+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/BuildModeHUD/Scripts/BuildModeHUD.asmdef
@@ -1,5 +1,6 @@
 {
   "name": "BuildModeHUD",
+  "rootNamespace": "",
   "references": [
     "GUID:555c1f3c6d18648df910b7a1de75b424",
     "GUID:98bf3bf29030cbf4092d89a7cc28b7fb",
@@ -24,7 +25,8 @@
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
     "GUID:28f74c468a54948bfa9f625c5d428f56",
     "GUID:bb3eef89092f71f44aa2449a9cf7879c",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
