+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Scene/Tests/SceneTests.asmdef
@@ -16,7 +16,8 @@
     "GUID:d7ced3c8f2800ed40a132e6df054755e",
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
     "GUID:b308ae946900fa04080222d7daf47d9a",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
