+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Environment/TestEnvironment/TestEnvironment.asmdef
@@ -16,7 +16,8 @@
         "GUID:1b07639381de24d0f932014911b5b207",
         "GUID:3d1a503a92d494d8aa3f35909695fa1f",
         "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:800ec6f2bae94e64098ee2c3a5e5d818"
+        "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+        "GUID:5289a39f46259954ca74010d7c9ded16"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
