+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a14cfaf4d3d47e849903378c7469d11a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
