+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Interfaces/IOutOfSceneBoundariesHandler.cs
@@ -1,0 +1,7 @@
+﻿namespace DCL.Controllers
+{
+    public interface IOutOfSceneBoundariesHandler
+    {
+        void UpdateOutOfBoundariesState(bool enable);
+    }
+}
