+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Interfaces/ISceneBoundsFeedbackStyle.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using DCL.Models;
+using UnityEngine;
+
+namespace DCL.Controllers
+{
+    public interface ISceneBoundsFeedbackStyle
+    {
+        void OnRendererExitBounds(Renderer renderer);
+        void ApplyFeedback(DCL.Models.MeshesInfo meshesInfo, bool isInsideBoundaries);
+        List<Material> GetOriginalMaterials(DCL.Models.MeshesInfo meshesInfo);
+    }
+}
