+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Interfaces/SceneBoundsCheckerInterfaces.asmdef
@@ -1,0 +1,17 @@
+{
+  "name": "SceneBoundsCheckerInterfaces",
+  "rootNamespace": "",
+  "references": [
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:8af21c3a612af2645a9953d4b9b3f6d5"
+  ],
+  "includePlatforms": [],
+  "excludePlatforms": [],
+  "allowUnsafeCode": false,
+  "overrideReferences": false,
+  "precompiledReferences": [],
+  "autoReferenced": true,
+  "defineConstraints": [],
+  "versionDefines": [],
+  "noEngineReferences": false
+}
