+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Interfaces/SceneBoundsCheckerInterfaces.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5289a39f46259954ca74010d7c9ded16
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
