+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/SceneBoundariesController/Tests/SceneBoundariesCheckerTests.asmdef
@@ -13,7 +13,8 @@
     "GUID:8af21c3a612af2645a9953d4b9b3f6d5",
     "GUID:1dd0780aa6be12b428c8005d0bee46b8",
     "GUID:b308ae946900fa04080222d7daf47d9a",
-    "GUID:97d8897529779cb49bebd400c7f402a6"
+    "GUID:97d8897529779cb49bebd400c7f402a6",
+    "GUID:5289a39f46259954ca74010d7c9ded16"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
