+++ unity-renderer/Assets/Scripts/MainScripts/MainScripts.asmdef
@@ -60,8 +60,8 @@
         "GUID:85ae2f0b8e8614c42bebf9817fc6f2be",
         "GUID:4e6796037298c492ea25360846f48fc3",
         "GUID:6e6a87f97192ec947993f360104e75b7",
-        "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:800ec6f2bae94e64098ee2c3a5e5d818"
+        "GUID:800ec6f2bae94e64098ee2c3a5e5d818",
+        "GUID:5289a39f46259954ca74010d7c9ded16"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
